+++ pkg/commands/compute/compute_test.go
@@ -101,8 +101,12 @@ func TestFlagDivergenceServe(t *testing.T) {
 	ignoreServeFlags := []string{
 		"addr",
 		"debug",
+		"experimental-enable-pushpin",
 		"file",
 		"profile-guest",
+		"pushpin-path",
+		"pushpin-proxy-port",
+		"pushpin-publish-port",
 		"profile-guest-dir",
 		"skip-build",
 		"viceroy-args",